pyproject.toml
# This file was generated from bswck/skeleton@61eeffb.
# Instead of changing this particular file, you might want to alter the template:
# https://github.com/bswck/skeleton/tree/61eeffb/project/pyproject.toml.jinja

[tool.poetry]
name = "redesc"
version = "0.0.0"
description = "A tool for substituting YouTube video descriptions with regex."
authors = ["bswck <bswck.dev@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "redesc/" }]
homepage = "https://github.com/apocomitamatma/redesc"

[tool.poetry.urls]
Issues = "https://github.com/apocomitamatma/redesc/issues"
Distribution = "https://pypi.org/project/redesc/"
Coverage = "https://coverage-badge.samuelcolvin.workers.dev/redirect/apocomitamatma/redesc"

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
configzen = "^0.9.0"
proxyvars = "^0.9.1"
hikari = "^2.0.0.dev120"
hikari-miru = "^3.1.2"
hikari-crescent = "^0.6.2"
google-api-python-client = "^2.89.0"
oauth2client = "<4.0.0"
google-auth-oauthlib = "^1.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]

[tool.poetry.group.dev-skeleton.dependencies]
# This dependency group was generated from bswck/skeleton@61eeffb.
# Instead of changing this particular file, you might want to alter the template:
# https://github.com/bswck/skeleton/tree/61eeffb/project/pyproject.toml.jinja
mypy = "^1.7.0"
ruff = "^0.1.2"
towncrier = "^23.11.0"
poethepoet = "^0.24.3"
pre-commit = "^3.5.0"

[tool.poe.tasks]
lint = "ruff check ."
skeleton = "scripts/skeleton.sh"
check = [
    { ref="lint" },
]
release.script = "scripts.release:main"

[tool.poe.tasks.added]
shell = "towncrier create $TICKET.added.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.changed]
shell = "towncrier create $TICKET.changed.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.fixed]
shell = "towncrier create $TICKET.fixed.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.deprecated]
shell = "towncrier create $TICKET.deprecated.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.removed]
shell = "towncrier create $TICKET.removed.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.poe.tasks.security]
shell = "towncrier create $TICKET.security.md --edit"
args = [{name = "TICKET", positional = true}]

[tool.ruff]
exclude = ["tests/", "scripts/"]
# To discuss the presented rationales, contact the author (bswck).
select = ["ALL"]
ignore = [
    # Description: Missing type annotation for self in method
    # Rationale: It is mostly redundant, as is for ANN102.
    "ANN101",

    # Description: Missing type annotation for cls in method
    # Rationale: It is mostly redundant, as is for ANN101.
    "ANN102",

    # Description: Dynamically typed expressions (typing.Any) are disallowed
    # Rationale:
    #  We use Any to allow user to constrain return types of our functions on their own.
    #  For example, having a function `def foo() -> Any: ...` allows user to write
    #  `my_foo: int = foo()` and have it passed type checking, since `Any` disables
    #  type checking completely.
    "ANN401",

    # Description: 1 blank line required before class docstring
    # Rationale: Remove the warning -- D211 (no-blank-line-before-class) preferred.
    "D203",

    # Description: Multi-line docstring summary should start at the second line
    # Rationale: Remove the warning -- D213 (multi-line-summary-second-line) preferred.
    "D212",

    # Description: Line contains TODO, consider resolving the issue
    # Rationale: Not appropriate for the project.
    "FIX002",

    "D",
    "A003",
    "DTZ003",
    "PLR",
    "PLW0603",
]

[tool.mypy]
strict = true

[tool.isort]
profile = "black"

[tool.towncrier]
directory = "changes"
package = "redesc"
filename = "CHANGELOG.md"
start_string = "<!-- insertion marker -->\n"
underlines = ["", "", ""]  # We use Markdown
title_format = "## [{version}](https://github.com/apocomitamatma/redesc/tree/{version}) ({project_date})"
issue_format = "[#{issue}](https://github.com/apocomitamatma/redesc/issues/{issue})"

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[tool.coverage.report]
omit = ["tests/*"]
# fail_under = 96
exclude_also = [
    # Be sure to include lines that are not meant to be covered
    # by tests or encountered by the users at runtime.
    "raise NotImplementedError",          # Do not cover abstract methods.
    "if ((t|typing)\\.?)?TYPE_CHECKING:", # Do not cover type checking imports.
    "@((t|typing)\\.?)?overload",         # Do not cover overload definitions.
    "\\(((t|typing)\\.?)?Protocol\\):$",  # Do not cover entire protocol definitions.
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
